changed the logger detection mechanism to play better with logging conventions
Previously the logger argument to the dog constructor could be either a string,
in which case it was looked up in the wrapped function's global scope,
or an object, in which case it would be used as-is.

Now if the argument is None (this is the default), the wrapped function
module's __name__ is passed to logging.getLogger() to get the logger.
If the argument is a string, that is sent to getLogger instead.
If the argument is any other object, that object is used as-is.